re-arrange test dir
